+++ Mythology_Parade_Engine/Core/CurrencySystem.h
@@ -5,6 +5,7 @@
 class CurrencySystem
 {
 public:
+	//ECONOMY BASICS
 	enum class ActionSacrifice {
 		RivalKilled,
 		EliteKilled,
@@ -21,17 +22,46 @@ public:
 		BeastConverted,
 		MonasteriesPrayers
 	};
+	//MIRACLES AND DISASTRES
+	enum class Miracles {
+		ClericEnlightment,
+		BordersPressure,
+		BeastTame,
+		MassConversion,
+		ForcedTraition,
+		WeakenedBorders,
+		CallToArms,
+		GoldenExpansion,
+		Victory
+	};
+	enum class Disasters {
+		MilitaryPressure,
+		TroopsInspiration,
+		HolyMeteor,
+		StrikingFear,
+		BloodFlow,
+		TheresRoom,
+		ColdAssassination,
+		AnimalShelter,
+		Victory
+	};
 
 public:
 	CurrencySystem();
 	~CurrencySystem();
 
+	//BASIC ECONOMY FUNCTONS
 	void increaseFaith(int number);
 	void decreaseFaith(int number);
 	void increaseSacrifice(ActionSacrifice action);
 	void increasePrayers(ActionPrayers action);
 
+	//MIRACLES AND DISASTER
+	void Miracle(Miracles action, CurrencySystem* currency);
+	void Disaster(Disasters action, CurrencySystem* currency);
+
 public:
+	//ECONOMY
 	int sacrifices;
 	int faith;
 	int prayers;
